# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WALAB API Documentation",
        "description": "Documentation technique complète pour l'application Walab. Utilisez le bouton 'Authorize' pour tester les routes protégées.",
        "contact": {
            "email": "contact@walab.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/commande/list": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Toutes les commandes",
                "description": "Récupère l'historique complet des commandes de l'utilisateur.",
                "operationId": "1773f41928780432e973f61b99b7c336",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Commande"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/list/pending": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Commandes en attente",
                "operationId": "2e4758d2f537415541e299f4b3159c61",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/list/progress": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Commandes en cours",
                "operationId": "3554499193e63e74551a7b732716688f",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/agent/list/pending": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Commandes Agent en attente",
                "operationId": "21b185c4bf9244bd6fb58bd3ced8cd06",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/agent/list/progress": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Commandes Agent en cours",
                "operationId": "09585daacbffdb2959395d3de287579f",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/list/finish": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Commandes terminées",
                "operationId": "b7fa241b4db4750f3e5b1751061a0755",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/agent/list/finish": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Commandes Agent terminées",
                "operationId": "a5886cb1bde9895d0812cd107baacd92",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/get": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Détail par ID",
                "operationId": "dce4de17b69aadd2808ea5e05f92a32e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/code": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Détail par Code",
                "operationId": "a9471d4d7d263f80005c1a608610c639",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/create": {
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Créer une commande",
                "operationId": "4e6e1ebcea203bd2ad1ecec8ebc24e7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "adress",
                                    "montant"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Prélèvement"
                                    },
                                    "adress": {
                                        "type": "string",
                                        "example": "Cotonou"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "payed": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "examen_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "type_bilan_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Créé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/update": {
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Mettre à jour une commande",
                "operationId": "284b7243fe21c762ae4792a92e86e75d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commande"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/delete": {
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Supprimer une commande",
                "operationId": "caef781d80c59f37cda2a10eb0c4192f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/update/statut": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Changer le statut",
                "operationId": "2bd6eaa183a0cf2f24837e3d7d47f342",
                "parameters": [
                    {
                        "name": "code_commande",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut changé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/admin/list": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Toutes les commandes (Admin)",
                "operationId": "6ea3ccdcd5b9fd36a43680fd32ba0e89",
                "responses": {
                    "200": {
                        "description": "Liste admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commande/agent/list": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Liste des commandes Agent",
                "operationId": "efc4200fa413c8d60506a2f7a0eff6c5",
                "responses": {
                    "200": {
                        "description": "Liste agent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Commande": {
                "title": "Commande",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "code": {
                        "type": "string",
                        "example": "CMD-8392"
                    },
                    "statut": {
                        "type": "string",
                        "example": "En cours"
                    },
                    "montant": {
                        "type": "number",
                        "example": 15000
                    },
                    "date_prelevement": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'accès ici (sans le préfixe 'Bearer')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}